# Description

This PR changes the implementation of `Distribute()` to use a dynamically generated select via the reflect package, instead of randomly choosing a single channel to send to.

In the case where all "out" channels are available to receive a message, the behavior is the same: the channel is selected pseudo-randomly.  However, the pseudo-random choice is made inside the Go library's `select` implementation, which is presumably highly optimized for this purpose.  `Benchmark_Distribute-8` performance is roughly doubled on my system (831 ns/op after the change vs. 1582 ns/op before).

In the case where some of the "out" channels are not able to receive a message right now, the `select` call will immediately choose one of the available channels.  If `Distribute()` is being used to send units of work to a pool of workers, we avoid the case where some workers sit idle because we happened to choose a busy worker to receive the next unit.

No changes are made to the library API and this should be fully backwards compatible with all library users, unless they rely on the exact details of the channel selection (and it's hard to see how one would do this).