@benjivesterby 2022-03-22
Benchmark update from my local test:

```shell
name            old time/op  new time/op  delta
_Distribute-16  2.86µs ± 2%  1.71µs ± 0%   ~     (p=0.400 n=4+1)
```